refactor(settings): extract OrchestrationSection (PR 4b — completes the cluster split)
Pull the orchestration parent section out of SettingsPanel into
OrchestrationSection.tsx, completing the split started in #586. The new
section owns its full state — `orchestrator`, `orchScope`, save tracker,
and `getOrchestratorSettings` load — and composes the three sub-components
(`PrMonitorSection`, `NotifySettingsSection`, `GuardrailsSection`) that
moved to their own files in the previous PR. The OrchestrationRuleTextarea
helper that lived next to SettingsPanel since #584 moves into the new file
since it has no other consumers.

After this PR, SettingsPanel.tsx is a thin layout shell over the section
components — no orchestrator state, no orchestrator-specific save tracker,
no refreshOrchestrator callback. The parent only passes `projects` (used
by the new section's scope selector).

- SettingsPanel.tsx: 759 → 530 lines (−229).
- Cumulative since #581: 2097 → 530, a 75% reduction.
- New file: OrchestrationSection.tsx (263 lines).
- 82/82 settings tests still pass; `pnpm build` clean; biome clean.
